--- a/common_description.go
+++ b/common_description.go
@@ -16,6 +16,10 @@ func (i Information) String() string {
 	return string(i)
 }
 
+func (i Information) marshalSize() (size int) {
+	return len(i)
+}
+
 // ConnectionInformation defines the representation for the "c=" field
 // containing connection data.
 type ConnectionInformation struct {
@@ -26,10 +30,20 @@ type ConnectionInformation struct {
 
 func (c ConnectionInformation) String() string {
 	parts := []string{c.NetworkType, c.AddressType}
-	if c.Address != nil && c.Address.String() != "" {
+	if c.Address != nil {
 		parts = append(parts, c.Address.String())
 	}
 	return strings.Join(parts, " ")
+}
+
+func (c ConnectionInformation) marshalSize() (size int) {
+	size = len(c.NetworkType)
+	size += 1 + len(c.AddressType)
+	if c.Address != nil {
+		size += 1 + c.Address.marshalSize()
+	}
+
+	return
 }
 
 // Address desribes a structured address token from within the "c=" field.
@@ -53,6 +67,18 @@ func (c *Address) String() string {
 	return strings.Join(parts, "/")
 }
 
+func (c Address) marshalSize() (size int) {
+	size = len(c.Address)
+	if c.TTL != nil {
+		size += 1 + lenUint(uint64(*c.TTL))
+	}
+	if c.Range != nil {
+		size += 1 + lenUint(uint64(*c.Range))
+	}
+
+	return
+}
+
 // Bandwidth describes an optional field which denotes the proposed bandwidth
 // to be used by the session or media.
 type Bandwidth struct {
@@ -70,11 +96,24 @@ func (b Bandwidth) String() string {
 	return output
 }
 
+func (b Bandwidth) marshalSize() (size int) {
+	if b.Experimental {
+		size += 2
+	}
+
+	size += len(b.Type) + 1 + lenUint(b.Bandwidth)
+	return
+}
+
 // EncryptionKey describes the "k=" which conveys encryption key information.
 type EncryptionKey string
 
-func (s EncryptionKey) String() string {
-	return string(s)
+func (e EncryptionKey) String() string {
+	return string(e)
+}
+
+func (e EncryptionKey) marshalSize() (size int) {
+	return len(e.String())
 }
 
 // Attribute describes the "a=" field which represents the primary means for
@@ -105,6 +144,15 @@ func (a Attribute) String() string {
 		output += ":" + a.Value
 	}
 	return output
+}
+
+func (a Attribute) marshalSize() (size int) {
+	size = len(a.Key)
+	if len(a.Value) > 0 {
+		size += 1 + len(a.Value)
+	}
+
+	return size
 }
 
 // IsICECandidate returns true if the attribute key equals "candidate".