--- a/time_description.go
+++ b/time_description.go
@@ -5,7 +5,6 @@ package sdp
 
 import (
 	"strconv"
-	"strings"
 )
 
 // TimeDescription describes "t=", "r=" fields of the session description
@@ -29,9 +28,12 @@ type Timing struct {
 }
 
 func (t Timing) String() string {
-	output := strconv.FormatUint(t.StartTime, 10)
-	output += " " + strconv.FormatUint(t.StopTime, 10)
-	return output
+	return stringFromMarshal(t.marshalInto, t.marshalSize)
+}
+
+func (t Timing) marshalInto(b []byte) []byte {
+	b = append(strconv.AppendUint(b, t.StartTime, 10), ' ')
+	return strconv.AppendUint(b, t.StopTime, 10)
 }
 
 func (t Timing) marshalSize() (size int) {
@@ -47,14 +49,19 @@ type RepeatTime struct {
 }
 
 func (r RepeatTime) String() string {
-	fields := make([]string, 0)
-	fields = append(fields, strconv.FormatInt(r.Interval, 10))
-	fields = append(fields, strconv.FormatInt(r.Duration, 10))
+	return stringFromMarshal(r.marshalInto, r.marshalSize)
+}
+
+func (r RepeatTime) marshalInto(b []byte) []byte {
+	b = strconv.AppendInt(b, r.Interval, 10)
+	b = append(b, ' ')
+	b = strconv.AppendInt(b, r.Duration, 10)
 	for _, value := range r.Offsets {
-		fields = append(fields, strconv.FormatInt(value, 10))
+		b = append(b, ' ')
+		b = strconv.AppendInt(b, value, 10)
 	}
 
-	return strings.Join(fields, " ")
+	return b
 }
 
 func (r RepeatTime) marshalSize() (size int) {