This reduces the amount of allocations need to marshal a SessionDescription

    Before
    '''
    BenchmarkMarshal-8        227802              5000 ns/op            2064 B/op         53 allocs/op
    BenchmarkMarshal-8        226938              5050 ns/op            2064 B/op         53 allocs/op
    BenchmarkMarshal-8        230073              5033 ns/op            2064 B/op         53 allocs/op
    BenchmarkMarshal-8        230949              5009 ns/op            2064 B/op         53 allocs/op
    BenchmarkMarshal-8        229460              4991 ns/op            2064 B/op         53 allocs/op
    '''

    After
    '''
    BenchmarkMarshal-8        460330              2237 ns/op             800 B/op          7 allocs/op
    BenchmarkMarshal-8        464844              2273 ns/op             800 B/op          7 allocs/op
    BenchmarkMarshal-8        475477              2271 ns/op             800 B/op          7 allocs/op
    BenchmarkMarshal-8        482371              2285 ns/op             800 B/op          7 allocs/op
    BenchmarkMarshal-8        479281              2265 ns/op             800 B/op          7 allocs/op
    '''